move the animation function to dpf_hdri_viewer.py
