+++ ansys_optical_automation/application/example_VRLab_animation.py
@@ -1,5 +1,3 @@
-import csv
-import time
 import tkinter as tk
 from tkinter import filedialog
 
@@ -31,59 +29,6 @@ def getfilename(extension, save=False):
     return file_path
 
 
-def animation_run(speos_file, csv_file):
-    """
-    function to run animation in observer result.
-
-    Parameters
-    ----------
-    speos_file : str
-        file path of VR file.
-    csv_file : str
-        file path of time line csv file
-
-    Returns
-    -------
-
-
-    """
-    VRLab = DpfHdriViewer()
-    VRLab.open_file(speos_file)
-    source_list = VRLab.get_source_list()
-    csv_source_list = []
-    csv_source_animation = []
-    with open(csv_file) as file:
-        content = csv.reader(file, delimiter=",")
-        First_Row = True
-        for row in content:
-            if First_Row:
-                csv_source_list = [item for item in row][1:]
-            else:
-                csv_source_animation.append([float(item) for item in row])
-            First_Row = False
-    animation_time_step = csv_source_animation[1][0] - csv_source_animation[0][0]
-    if len(csv_source_list) != len(source_list):
-        msg = "selected timeline csv file does not match with the speos vr file"
-        raise ImportError(msg)
-
-    if set(csv_source_list) != set(source_list):
-        print("will assume source using index")
-        VRLab.dpf_instance.Show(1)
-        while True:
-            for csv_source_config in csv_source_animation:
-                for source_idx, source in enumerate(csv_source_list):
-                    VRLab.set_source_power(source_idx, csv_source_config[source_idx + 1])
-                time.sleep(animation_time_step)
-    else:
-        print("will use name to set source in animation")
-        VRLab.dpf_instance.Show(1)
-        while True:
-            for csv_source_config in csv_source_animation:
-                for source_idx, source in enumerate(csv_source_list):
-                    VRLab.set_source_power(source, csv_source_config[source_idx + 1])
-                time.sleep(animation_time_step)
-
-
 def main():
     """
     main function to run Observer VR animation.
@@ -99,7 +44,9 @@ def main():
     if not vr_file or not time_line_csv:
         msg = "Please select speos vr file and time line csv file"
         raise ImportError(msg)
-    animation_run(vr_file, time_line_csv)
+    VRLab = DpfHdriViewer()
+    VRLab.open_file(vr_file)
+    VRLab.timeline_animation_run(time_line_csv)
 
 
 main()
